# modules/apps/headless/headless-admin-taxonomy/headless-admin-taxonomy-impl/rest-openapi.yaml
components:
    schemas:
...
        Creator:
            description:
                Represents the user who created the content. Properties follow the [creator](https://schema.org/creator)
                specification.
            properties:
                additionalName:
                    description:
                        An additional name for the user. This can be used for a middle name.
                    readOnly: true
                    type: string
                contentType:
                    # @review
                    description:
                        The type of the content.
                    readOnly: true
                    type: string
                familyName:
                    description:
                        The user's surname (last name).
                    readOnly: true
                    type: string
                givenName:
                    description:
                        The user's first name.
                    readOnly: true
                    type: string
                id:
                    description:
                        The user's ID.
                    format: int64
                    readOnly: true
                    type: integer
                image:
                    description:
                        A relative URL to the user's profile image.
                    format: uri
                    readOnly: true
                    type: string
                name:
                    description:
                        The user's full name.
                    readOnly: true
                    type: string
                profileURL:
                    description:
                        A relative URL to the user's profile.
                    format: uri
                    readOnly: true
                    type: string
            type: object
...
        TaxonomyCategory:
            description:
                Represents a category, which is a hierarchical classification that can be associated with particular
                asset types. Properties follow the [category](https://schema.org/category) specification.
            properties:
                availableLanguages:
                    description:
                        A list of languages the category has a translation for.
                    items:
                        type: string
                    readOnly: true
                    type: array
                creator:
                    allOf:
                        - $ref: "#/components/schemas/Creator"
                    description:
                        The category's creator.
                    readOnly: true
                dateCreated:
                    description:
                        The category's creation date.
                    format: date-time
                    readOnly: true
                    type: string
                dateModified:
                    description:
                        The category's most recent modification date.
                    format: date-time
                    readOnly: true
                    type: string
                description:
                    description:
                        The category's text description.
                    type: string
                externalReferenceCode:
                    type: string
                id:
                    description:
                        The category's ID.
                    readOnly: true
                    type: string
                name:
                    description:
                        The category's name.
                    type: string
                numberOfTaxonomyCategories:
                    description:
                        The number of times this category has been used in other assets.
                    readOnly: true
                    type: integer
                parentTaxonomyCategory:
                    description:
                        The category's parent category, if it exists.
                    properties:
                        id:
                            format: int64
                            type: integer
                        name:
                            type: string
                    readOnly: true
                    type: object
                parentTaxonomyVocabulary:
                    description:
                        The parent category's `TaxonomyVocabulary`, if such a parent category exists.
                    properties:
                        id:
                            format: int64
                            type: integer
                        name:
                            type: string
                    readOnly: true
                    type: object
                viewableBy:
                    description:
                        A write-only property that specifies the category's default permissions.
                    enum: [Anyone, Members, Owner]
                    type: string
                    writeOnly: true
            required:
                - name
            type: object
...
info:
    description:
        ""
    license:
        name: "Apache 2.0"
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
    title: "Headless Admin Taxonomy"
    version: v1.0
openapi: 3.0.1
paths:
...
    "/taxonomy-categories/{parentTaxonomyCategoryId}/taxonomy-categories":
        get:
            description:
                Retrieves a taxonomy category's child taxonomy categories. Results can be paginated, filtered, searched,
                and sorted.
            operationId: getTaxonomyCategoryTaxonomyCategoriesPage
            parameters:
                - in: path
                  name: parentTaxonomyCategoryId
                  required: true
                  schema:
                      type: string
                - in: query
                  name: filter
                  schema:
                      type: string
                - in: query
                  name: page
                  schema:
                      type: integer
                - in: query
                  name: pageSize
                  schema:
                      type: integer
                - in: query
                  name: search
                  schema:
                      type: string
                - in: query
                  name: sort
                  schema:
                      type: string
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: "#/components/schemas/TaxonomyCategory"
                                type: array
                        application/xml:
                            schema:
                                items:
                                    $ref: "#/components/schemas/TaxonomyCategory"
                                type: array
                    description:
                        ""
            tags: ["TaxonomyCategory"]
        post:
            description:
                Inserts a new child taxonomy category.
            operationId: postTaxonomyCategoryTaxonomyCategory
            parameters:
                - in: path
                  name: parentTaxonomyCategoryId
                  required: true
                  schema:
                      type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/TaxonomyCategory"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/TaxonomyCategory"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/TaxonomyCategory"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/TaxonomyCategory"
                    description:
                        ""
            tags: ["TaxonomyCategory"]
    "/taxonomy-categories/{taxonomyCategoryId}":
        delete:
            description:
                Deletes the taxonomy category and returns a 204 if the operation succeeds.
            operationId: deleteTaxonomyCategory
            parameters:
                - in: path
                  name: taxonomyCategoryId
                  required: true
                  schema:
                      type: string
            responses:
                204:
                    content:
                        application/json: {}
                        application/xml: {}
                    description:
                        ""
            tags: ["TaxonomyCategory"]
        get:
            description:
                Retrieves a taxonomy category.
            operationId: getTaxonomyCategory
            parameters:
                - in: path
                  name: taxonomyCategoryId
                  required: true
                  schema:
                      type: string
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/TaxonomyCategory"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/TaxonomyCategory"
                    description:
                        ""
            tags: ["TaxonomyCategory"]
        patch:
            description:
                Updates only the fields received in the request body. Other fields are left untouched.
            operationId: patchTaxonomyCategory
            parameters:
                - in: path
                  name: taxonomyCategoryId
                  required: true
                  schema:
                      type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/TaxonomyCategory"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/TaxonomyCategory"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/TaxonomyCategory"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/TaxonomyCategory"
                    description:
                        ""
            tags: ["TaxonomyCategory"]
        put:
            description:
                Replaces the taxonomy category with the information sent in the request body. Any missing fields are
                deleted unless they are required.
            operationId: putTaxonomyCategory
            parameters:
                - in: path
                  name: taxonomyCategoryId
                  required: true
                  schema:
                      type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/TaxonomyCategory"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/TaxonomyCategory"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/TaxonomyCategory"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/TaxonomyCategory"
                    description:
                        ""
            tags: ["TaxonomyCategory"]
...
    "/taxonomy-vocabularies/{taxonomyVocabularyId}/taxonomy-categories":
        get:
            description:
                Retrieves a vocabulary's taxonomy categories. Results can be paginated, filtered, searched, and sorted.
            operationId: getTaxonomyVocabularyTaxonomyCategoriesPage
            parameters:
                - in: path
                  name: taxonomyVocabularyId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: query
                  name: filter
                  schema:
                      type: string
                - in: query
                  name: page
                  schema:
                      type: integer
                - in: query
                  name: pageSize
                  schema:
                      type: integer
                - in: query
                  name: search
                  schema:
                      type: string
                - in: query
                  name: sort
                  schema:
                      type: string
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: "#/components/schemas/TaxonomyCategory"
                                type: array
                        application/xml:
                            schema:
                                items:
                                    $ref: "#/components/schemas/TaxonomyCategory"
                                type: array
                    description:
                        ""
            tags: ["TaxonomyCategory"]
        post:
            description:
                Inserts a new taxonomy category in a taxonomy vocabulary.
            operationId: postTaxonomyVocabularyTaxonomyCategory
            parameters:
                - in: path
                  name: taxonomyVocabularyId
                  required: true
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/TaxonomyCategory"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/TaxonomyCategory"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/TaxonomyCategory"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/TaxonomyCategory"
                    description:
                        ""
            tags: ["TaxonomyCategory"]